// frontend/src/openapi/compassClient/models/TimestampDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TimestampDto
 */
export interface TimestampDto {
    /**
     * 
     * @type {number}
     * @memberof TimestampDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof TimestampDto
     */
    daySheetId?: number;
    /**
     * 
     * @type {string}
     * @memberof TimestampDto
     */
    startTime?: string;
    /**
     * 
     * @type {string}
     * @memberof TimestampDto
     */
    endTime?: string;
}

/**
 * Check if a given object implements the TimestampDto interface.
 */
export function instanceOfTimestampDto(value: object): boolean {
    return true;
}

export function TimestampDtoFromJSON(json: any): TimestampDto {
    return TimestampDtoFromJSONTyped(json, false);
}

export function TimestampDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimestampDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'daySheetId': json['day_sheet_id'] == null ? undefined : json['day_sheet_id'],
        'startTime': json['start_time'] == null ? undefined : json['start_time'],
        'endTime': json['end_time'] == null ? undefined : json['end_time'],
    };
}

export function TimestampDtoToJSON(value?: TimestampDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'day_sheet_id': value['daySheetId'],
        'start_time': value['startTime'],
        'end_time': value['endTime'],
    };
}